requirements.txt
google-genai>=0.2.2
pydantic>=2.6.3
python-dotenv>=1.0.1
